Shift ScalarType et al. to pvType.h
Allows shared_vector and AnyScalar to
avoid including pvIintrospect.h
