+++ src/misc/anyscalar.cpp
@@ -1,4 +1,5 @@
 
+#include <string.h>
 #include <epicsAssert.h>
 
 #define epicsExportSharedSymbols
